searching is completed w.r.t logged in user

+++ app/containers/HomePage/userActionCreator.js
@@ -2,33 +2,33 @@ import { toast } from 'react-toastify';
 
 import * as actionTypes from './userActionTypes';
 
-export const searchWordAction = (word, searchType, synonym) => (dispatch, getState, { api }) => {
+export const searchWordAction = (word, searchType, synonym, userId) => (dispatch, getState, { api }) => {
   dispatch({
     type: actionTypes.SEARCH_WORD,
     payload: { searchText: word, searchType, isSynonym: !!synonym },
-    promise: api.get(`/api/searchWord/${encodeURI(word)}/${searchType}`).then((res) => res.data, () => {
+    promise: api.get(`/api/searchWord/${encodeURI(word)}/${searchType}/${userId}`).then((res) => res.data, () => {
       toast.error('Something went wrong. Please try again!');
       throw new Error();
     })
   });
 };
 
-export const loadWordAction = () => (dispatch, getState, { api }) => {
+export const loadWordAction = (userId) => (dispatch, getState, { api }) => {
   dispatch({
     type: actionTypes.LOAD_WORDS,
     payload: {},
-    promise: api.get('/api/getAllWords').then((res) => res.data, () => {
+    promise: api.get(`/api/getAllWords/${userId}`).then((res) => res.data, () => {
       toast.error('Something went wrong. Please try again!');
       throw new Error();
     })
   });
 };
 
-export const addWordAction = (wordDetails) => (dispatch, getState, { api }) => {
+export const addWordAction = (wordDetails, userId) => (dispatch, getState, { api }) => {
   dispatch({
     type: actionTypes.ADD_WORD,
     payload: {},
-    promise: api.post('/api/addWordToCollection', wordDetails).then((res) => {
+    promise: api.post('/api/addWordToCollection', { ...wordDetails, userId }).then((res) => {
       toast.success(`${wordDetails.word} has been bookmarked."`);
       return res.data;
     }, () => {
@@ -38,11 +38,11 @@ export const addWordAction = (wordDetails) => (dispatch, getState, { api }) => {
   });
 };
 
-export const deleteWordAction = (word) => (dispatch, getState, { api }) => {
+export const deleteWordAction = (word, userId) => (dispatch, getState, { api }) => {
   dispatch({
     type: actionTypes.DELETE_WORD,
     payload: { word },
-    promise: api.delete(`/api/deleteWord/${encodeURI(word)}`).then((res) => {
+    promise: api.delete(`/api/deleteWord/${encodeURI(word)}`, { userId }).then((res) => {
       toast.success(`${word} has been removed from bookmark."`);
       return res.data;
     }, () => {
